# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ShopNow E-Commerce API
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  message: string;
  errors?: string[];
}

export interface MessageResponse {
  message: string;
}

export interface RegisterRequest {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  user: "user",
  admin: "admin",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  avatar?: string | null;
  address?: string | null;
  phone?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UpdateUserRequest {
  name?: string;
  avatar?: string | null;
  address?: string | null;
  phone?: string | null;
}

export interface UsersResponse {
  users: User[];
  total: number;
  page: number;
  totalPages: number;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  originalPrice?: number | null;
  category: string;
  image: string;
  images: string[];
  stock: number;
  rating: number;
  numReviews: number;
  featured: boolean;
  tags: string[];
  createdAt: string;
}

export interface CreateProductRequest {
  name: string;
  description: string;
  price: number;
  originalPrice?: number | null;
  category: string;
  image: string;
  images?: string[];
  stock: number;
  featured?: boolean;
  tags?: string[];
}

export interface ProductsResponse {
  products: Product[];
  total: number;
  page: number;
  totalPages: number;
}

export interface CategoriesResponse {
  categories: string[];
}

export interface OrderItem {
  productId: number;
  name: string;
  image: string;
  price: number;
  quantity: number;
}

export interface ShippingAddress {
  fullName: string;
  address: string;
  city: string;
  postalCode: string;
  country: string;
}

export type OrderPaymentStatus =
  (typeof OrderPaymentStatus)[keyof typeof OrderPaymentStatus];

export const OrderPaymentStatus = {
  pending: "pending",
  paid: "paid",
  failed: "failed",
  refunded: "refunded",
} as const;

export type OrderOrderStatus =
  (typeof OrderOrderStatus)[keyof typeof OrderOrderStatus];

export const OrderOrderStatus = {
  pending: "pending",
  processing: "processing",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface Order {
  id: number;
  userId: number;
  items: OrderItem[];
  shippingAddress: ShippingAddress;
  paymentMethod: string;
  paymentStatus: OrderPaymentStatus;
  orderStatus: OrderOrderStatus;
  subtotal: number;
  shippingCost: number;
  tax: number;
  total: number;
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
  user?: User | null;
}

export type CreateOrderRequestItemsItem = {
  productId: number;
  quantity: number;
};

export interface CreateOrderRequest {
  items: CreateOrderRequestItemsItem[];
  shippingAddress: ShippingAddress;
  paymentMethod: string;
  notes?: string | null;
}

export type UpdateOrderStatusRequestOrderStatus =
  (typeof UpdateOrderStatusRequestOrderStatus)[keyof typeof UpdateOrderStatusRequestOrderStatus];

export const UpdateOrderStatusRequestOrderStatus = {
  pending: "pending",
  processing: "processing",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export type UpdateOrderStatusRequestPaymentStatus =
  (typeof UpdateOrderStatusRequestPaymentStatus)[keyof typeof UpdateOrderStatusRequestPaymentStatus];

export const UpdateOrderStatusRequestPaymentStatus = {
  pending: "pending",
  paid: "paid",
  failed: "failed",
  refunded: "refunded",
} as const;

export interface UpdateOrderStatusRequest {
  orderStatus: UpdateOrderStatusRequestOrderStatus;
  paymentStatus?: UpdateOrderStatusRequestPaymentStatus;
}

export interface OrdersResponse {
  orders: Order[];
  total: number;
  page: number;
  totalPages: number;
}

export interface WishlistItem {
  id: number;
  productId: number;
  product: Product;
  createdAt: string;
}

export interface WishlistResponse {
  items: WishlistItem[];
}

export interface WishlistRequest {
  productId: number;
}

export type AdminStatsOrdersByStatus = { [key: string]: number };

export type AdminStatsRevenueByMonthItem = {
  month: string;
  revenue: number;
};

export interface AdminStats {
  totalRevenue: number;
  totalOrders: number;
  totalProducts: number;
  totalUsers: number;
  recentOrders: Order[];
  ordersByStatus: AdminStatsOrdersByStatus;
  revenueByMonth: AdminStatsRevenueByMonthItem[];
}

export type GetProductsParams = {
  search?: string;
  category?: string;
  minPrice?: number;
  maxPrice?: number;
  sortBy?: GetProductsSortBy;
  page?: number;
  limit?: number;
  inStock?: boolean;
};

export type GetProductsSortBy =
  (typeof GetProductsSortBy)[keyof typeof GetProductsSortBy];

export const GetProductsSortBy = {
  price_asc: "price_asc",
  price_desc: "price_desc",
  name_asc: "name_asc",
  name_desc: "name_desc",
  newest: "newest",
  rating: "rating",
} as const;

export type GetUsersParams = {
  page?: number;
  limit?: number;
};

export type GetOrdersParams = {
  page?: number;
  limit?: number;
  status?: GetOrdersStatus;
};

export type GetOrdersStatus =
  (typeof GetOrdersStatus)[keyof typeof GetOrdersStatus];

export const GetOrdersStatus = {
  pending: "pending",
  processing: "processing",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;