# lib/api-zod/src/generated/types/createProductRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ShopNow E-Commerce API
 * OpenAPI spec version: 1.0.0
 */

export interface CreateProductRequest {
  name: string;
  description: string;
  price: number;
  originalPrice?: number | null;
  category: string;
  image: string;
  images?: string[];
  stock: number;
  featured?: boolean;
  tags?: string[];
}