# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ShopNow E-Commerce API
 * OpenAPI spec version: 1.0.0
 */
import type { OrderItem } from "./orderItem";
import type { OrderOrderStatus } from "./orderOrderStatus";
import type { OrderPaymentStatus } from "./orderPaymentStatus";
import type { ShippingAddress } from "./shippingAddress";
import type { User } from "./user";

export interface Order {
  id: number;
  userId: number;
  items: OrderItem[];
  shippingAddress: ShippingAddress;
  paymentMethod: string;
  paymentStatus: OrderPaymentStatus;
  orderStatus: OrderOrderStatus;
  subtotal: number;
  shippingCost: number;
  tax: number;
  total: number;
  notes?: string | null;
  createdAt: Date;
  updatedAt: Date;
  user?: User | null;
}