# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ShopNow E-Commerce API
 * OpenAPI spec version: 1.0.0
 */

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  originalPrice?: number | null;
  category: string;
  image: string;
  images: string[];
  stock: number;
  rating: number;
  numReviews: number;
  featured: boolean;
  tags: string[];
  createdAt: Date;
}